requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.4
idna==3.10
pillow==11.1.0
python-dateutil==2.9.0.post0
razorpay==1.4.2
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.4.0
